/runs api working

+++ mip/server/runs_api.py
@@ -1,5 +1,6 @@
 # Copyright 2024 InferLink Corporation
 import json
+import pdb
 from datetime import datetime
 from pathlib import Path
 import subprocess
@@ -15,19 +16,23 @@ class RunsApi:
         self._threads: list[Thread] = []
         return
 
-    def start_run(self, body: RunPayloadModel) -> RunStatusModel:
+    def post_run(self, body: RunPayloadModel) -> RunStatusModel:
         run_id = self._create_run_id()
+
+        filename = self._get_run_id_filename(run_id)
+        if filename.exists():
+            # TODO: don't run more than once a second...
+            raise NotImplementedError("too many runs at the same time")
 
         run_status = RunStatusModel(
             run_id=run_id,
             status=Status.RUNNING,
             payload=body,
             start_time=datetime.now(),
+            stop_time=None,
             end_time=None,
             log=None,
         )
-
-        filename = self._get_run_id_filename(run_id)
         filename.write_text(run_status.model_dump_json())
 
         t = Thread(target=self._run_mipper, args=(run_status,))
@@ -38,7 +43,7 @@ class RunsApi:
         return run_status
 
     def _run_mipper(self, run_status: RunStatusModel) -> None:
-        stat = subprocess.run(args=[], capture_output=True, stderr=subprocess.STDOUT, text=True)
+        stat = subprocess.run(args=["ls"], capture_output=True, text=True)  # stderr=subprocess.STDOUT
 
         if stat.returncode == 0:
             run_status.status = Status.PASSED
@@ -47,16 +52,16 @@ class RunsApi:
 
         run_status.log = stat.stdout
 
-        run_status.payload = datetime.now()
+        run_status.stop_time = datetime.now()
 
         filename = self._get_run_id_filename(run_status.run_id)
         filename.write_text(run_status.model_dump_json())
 
-    def get_run_ids(self) -> list[str]:
+    def get_runs(self) -> list[str]:
         files = self._configuration.host.runs_dir.glob("*.json")
         return [f.stem for f in files]
 
-    def get_run_status(self, run_id: str) -> RunStatusModel:
+    def get_run_by_name(self, run_id: str) -> RunStatusModel:
         filename = self._get_run_id_filename(run_id)
         data = json.loads(filename.read_text())
         run_status = RunStatusModel(**data)
@@ -64,7 +69,7 @@ class RunsApi:
 
     @staticmethod
     def _create_run_id() -> str:
-        s = datetime.now().strftime("%Y%m%d-%H%M%S-%f")
+        s = datetime.now().strftime("%Y%m%d-%H%M%S")
         return s
 
     def _get_run_id_filename(self, run_id: str) -> Path:
